내용 : 2단계 - 로또(수동) - 코드리뷰 요청 2차 이슈 : -

+++ src/main/java/lotto/view/LottoOutput.java
@@ -25,7 +25,7 @@ public class LottoOutput {
     }
 
     public void displayPriceEarningsRatio(BuyLotto buyLotto) {
-        float priceEarninsRatio = ((EqualNumbersMoney.EQUALS.getWinAmountTotal() - buyLotto.getbuyAmount()) / buyLotto.getbuyAmount()) * 100;
+        float priceEarninsRatio = ((EqualNumbersMoney.EQUALS.getWinAmountTotal() - buyLotto.getbuyAmount()) / (float) buyLotto.getbuyAmount()) * 100;
 
         System.out.println("총 수익률은 " + priceEarninsRatio + "% 입니다.");
     }
